Created an IE shim to deal with the fact that I.E. does not support "new window.URL" Continuing to figure out the bugs in alasql.  The main
Created an IE shim to deal with the fact that I.E. does not support "new
window.URL"
Continuing to figure out the bugs in alasql.  The main issue right now
is that the database is not attaching itself to the member and txs
tables so that they get lost somehow.  This could be due to the
asynchronous nature of alasql, but the "when/then" promise functionality
should be working.